[PRDI-602] Delete Pending Files On Network Failure (#1)
When members use a tool like Little Snitch or pi-hole to block outbound requests to segment.io, the Segment library will persist these analytics to disk into different files. These files can stack up immensely over time. This could mean you are trying to make 1,000s of network calls to upload these analytics that will never succeed. We would prefer to give this CPU time back to members at the trade off of maybe having some fewer events for the time being.

This means that as events fail to upload we will slowly delete the temporary files on disk so that they are not building up over time. This means that we will likely give up some analytics event integrity to give back CPU time and general performance to member machines that are in this state (i.e. people using this kind of blocking software).